requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openpyxl
pdfplumber
python-multipart
httpx
anthropic
python-dotenv
apscheduler
aiofiles
google-genai
